--- a/Dockerfile.bench
+++ b/Dockerfile.bench
@@ -1,0 +1,51 @@
+FROM debian:buster-slim
+
+# Prerequisites of Scala Native
+RUN set -x \
+  && apt-get update \
+  && apt-get install -y clang libre2-dev libunwind-dev \
+  && rm -rf /var/lib/apt/lists/*
+
+
+RUN set -x \
+  `# workaround for error in debian-slim during openjdk installation` \
+  && mkdir -p /usr/share/man/man1 \
+  && apt-get update \
+  && apt-get install --no-install-recommends -y curl git gnupg openjdk-11-jre-headless python3 \
+  && rm -rf /var/lib/apt/lists/*
+
+RUN set -x \
+  && echo "deb https://dl.bintray.com/sbt/debian /" | tee -a /etc/apt/sources.list.d/sbt.list \
+  && curl -fsL "https://keyserver.ubuntu.com/pks/lookup?op=get&search=0x2EE0EA64E40A89B84B2DF73499E82A75642AC823" | apt-key add \
+  && apt-get update \
+  && apt-get install -y sbt \
+  && rm -rf /var/lib/apt/lists/*
+
+RUN set -x \
+ && apt-get update \
+ && apt-get install --no-install-recommends -y build-essential python3-dev python3-pip \
+ && rm -rf /var/lib/apt/lists/*
+# https://stackoverflow.com/questions/52904639/dockerfile-with-pip-grpcio-is-very-slow-to-build => fix not applied yet
+
+RUN set -x \
+  && apt-get update \
+  && apt-get install -y openjdk-11-jdk-headless \
+  && rm -rf /var/lib/apt/lists/*
+
+RUN set -x \
+  && apt-get update \
+  && apt-get install -y apt-file \
+  && apt-file update \
+  && apt-get install -y vim \
+  && apt-get install -y wget
+
+RUN set -x \
+ && wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda.sh \
+ && bash ~/miniconda.sh -b -p ~/miniconda \
+ && . ~/miniconda/bin/activate \
+ && conda install --yes --quiet python=3.7 \
+ && conda install --yes --quiet tensorflow=2.2.0 \
+ && conda install --yes --quiet numpy=1.18.5
+
+RUN set -x \
+ && echo ". ~/miniconda/bin/activate" >> /etc/bash.bashrc