--- a/Dockerfile.bench
+++ b/Dockerfile.bench
@@ -9,7 +9,6 @@ RUN set -x \
 RUN set -x \
   `# workaround for error in debian-slim during openjdk installation` \
   && mkdir -p /usr/share/man/man1 \
-  && apt-get update \
   && apt-get install --no-install-recommends -y curl git gnupg openjdk-11-jre-headless python3
 
 RUN set -x \
@@ -19,12 +18,10 @@ RUN set -x \
   && apt-get install -y sbt
 
 RUN set -x \
- && apt-get update \
  && apt-get install --no-install-recommends -y build-essential python3-dev python3-pip
 
 RUN set -x \
-  && apt-get update \
-  && apt-get install -y vim wget
+  && apt-get install --no-install-recommends -y vim wget
 
 RUN set -x \
  && wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda.sh \