… Dockerfile.base
… Dockerfile.conda
… Dockerfile.pip