as in issue #43 
Testing done on `BidirectionalLSTMExample` because it's a lot faster then MNIST.
Added python version with identical configuration and a Docker file that can be used as testing environment.
Results:
SCALA:
Total Time: 108.5
Epoch 1 Time: 46s 59ms/step 
Epoch 2 Time: 41s 53ms/step 
Test loss: 0.638
Test accuracy: 0.619

PYTHON:
Total Time: 107.8
Epoch 1 Time: 41s 52ms/step 
Epoch 2 Time: 48s 61ms/step 
Test loss: 0.643
Test accuracy: 0.615

So the performance is more or less identical.